Update project2.java
I checked mine and Ian's code and took pieces from each and joined it together, also cleaning it u in the process to make it more neat, I'm pretty sure it all works up until you choose a topic, if you find anything wrong then let me know. 

All we need now is the text file with all the questions and stuff in it, and then we just output it to the user.